fix(sidebar): address PR review comments on share modal stack
- Fix snackbar label having literal quotes around text in edit-share-modal
- Fix refreshGrants to clear to [] when server returns no grants (last revoke stale)
- Guard against missing grant.zid in ShareRevokeModal before issuing API call
- Wrap sendShareNotificationSoapApi in try/catch so network errors don't block revoke
- Fix trailing space in edit-share-modal title
- Fix pemissions typo in share note text (add-share-modal, edit-share-modal)
- Split edit-permissions-modal.test.tsx into add-share-modal.test.tsx and edit-share-modal.test.tsx
