+++ src/views/sidebar/add-share-modal.tsx
@@ -178,7 +178,7 @@ export const AddShareModal: FC<AddShareModalProps> = ({ onClose, folder, goBack,
 						<Text overflow="break-word" size="small" color="gray1">
 							{t(
 								'share.share_note',
-								'The standard message displays your name, the name of the shared item, pemissions granted to the recipients, and sign in information, if necessary.'
+								'The standard message displays your name, the name of the shared item, permissions granted to the recipients, and sign in information, if necessary.'
 							)}
 						</Text>
 					</Row>
