+++ src/views/sidebar/parts/edit/tests/share-folder-properties.test.tsx
@@ -184,5 +184,4 @@ describe('ShareFolderProperties', () => {
 			expect(chip).toHaveClass(defaultClass, { exact: true });
 		});
 	});
-
 });
