+++ src/views/sidebar/tests/share-revoke-modal.test.tsx
@@ -178,5 +178,4 @@ describe('ShareRevokeModal', () => {
 			expect(onSuccess).not.toHaveBeenCalled();
 		});
 	});
-
 });
